make product.notification_threshold integer

+++ bot_common/database/models/products.py
@@ -15,7 +15,7 @@ class Product(base.Base, mixin.CreatedMixin):
 
     id: Mapped[int] = mapped_column(primary_key=True)
     link: Mapped[str] = mapped_column(index=True)
-    notification_threshold: Mapped[float]
+    notification_threshold: Mapped[int]
     initial_price: Mapped[str]
     last_checked_at: Mapped[dt.datetime | None]
